Separate files for CScriptID and CTxDestination.
