+++ _bitcoin_common.vcxproj.filters
@@ -42,9 +42,6 @@
     <ClCompile Include="src\protocol.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\pubkey.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="src\random.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -325,6 +322,15 @@
     <ClCompile Include="src\script\standard.cpp">
       <Filter>Source Files\key</Filter>
     </ClCompile>
+    <ClCompile Include="src\TxDestination.cpp">
+      <Filter>Source Files\key</Filter>
+    </ClCompile>
+    <ClCompile Include="src\ScriptID.cpp">
+      <Filter>Source Files\key</Filter>
+    </ClCompile>
+    <ClCompile Include="src\pubkey.cpp">
+      <Filter>Source Files\key</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\addrdb.h">
@@ -510,9 +516,6 @@
     <ClInclude Include="src\protocol.h">
       <Filter>Source Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\pubkey.h">
-      <Filter>Source Files</Filter>
-    </ClInclude>
     <ClInclude Include="src\random.h">
       <Filter>Source Files</Filter>
     </ClInclude>
@@ -692,6 +695,15 @@
       <Filter>Source Files\key</Filter>
     </ClInclude>
     <ClInclude Include="src\script\standard.h">
+      <Filter>Source Files\key</Filter>
+    </ClInclude>
+    <ClInclude Include="src\TxDestination.h">
+      <Filter>Source Files\key</Filter>
+    </ClInclude>
+    <ClInclude Include="src\ScriptID.h">
+      <Filter>Source Files\key</Filter>
+    </ClInclude>
+    <ClInclude Include="src\pubkey.h">
       <Filter>Source Files\key</Filter>
     </ClInclude>
   </ItemGroup>
