+++ src/script/standard.h
@@ -12,86 +12,17 @@
 #include "script/interpreter.h"
 #include "uint256.h"
 #include "pubkey.h" // CKeyID
+#include "ScriptID.h"
+#include "TxDestination.h"
 
 static const bool DEFAULT_ACCEPT_DATACARRIER = true;
 
-// class CKeyID;
 class CScript;
-class base58string;
 
-/** A reference to a CScript: the Hash160 of its serialization (see script.h) */
-class CScriptID : public uint160 {
-public:
-    CScriptID() : uint160() {}
-    CScriptID(const CScript& in);
-    CScriptID(const uint160& in) : uint160(in) {}
-public:
-    base58string ToBase58address() const;
-};
 
 static const unsigned int MAX_OP_RETURN_RELAY = 83; //!< bytes (+1 for OP_RETURN, +2 for the pushdata opcodes)
 extern bool fAcceptDatacarrier;
 extern unsigned nMaxDatacarrierBytes;
-
-/**
- * Mandatory script verification flags that all new blocks must comply with for
- * them to be valid. (but old blocks may not comply with) Currently just P2SH,
- * but in the future other flags may be added, such as a soft-fork to enforce
- * strict DER encoding.
- * 
- * Failing one of these tests may trigger a DoS ban - see CheckInputs() for
- * details.
- */
-static const unsigned int MANDATORY_SCRIPT_VERIFY_FLAGS = SCRIPT_VERIFY_P2SH;
-
-enum txnouttype
-{
-    TX_NONSTANDARD,
-    // 'standard' transaction types:
-    TX_PUBKEY,
-    TX_PUBKEYHASH,
-    TX_SCRIPTHASH,
-    TX_MULTISIG,
-    TX_NULL_DATA,
-    TX_WITNESS_V0_SCRIPTHASH,
-    TX_WITNESS_V0_KEYHASH,
-};
-
-class CNoDestination {
-public:
-    friend bool operator==(const CNoDestination &a, const CNoDestination &b) { return true; }
-    friend bool operator<(const CNoDestination &a, const CNoDestination &b) { return true; }
-};
-
-/** 
- * A txout script template with a specific destination. It is either:
- *  * CNoDestination: no destination set
- *  * CKeyID: TX_PUBKEYHASH destination
- *  * CScriptID: TX_SCRIPTHASH destination
- *  A CTxDestination is the internal data type encoded in a CBitcoinAddress
- */
-class CTxDestination : public boost::variant<CNoDestination, CKeyID, CScriptID> {
-public:
-    typedef boost::variant<CNoDestination, CKeyID, CScriptID> Super;
-public:
-    CTxDestination() {}
-    CTxDestination(const CNoDestination& rhs) : Super(rhs) {}
-    CTxDestination(const CKeyID& rhs) : Super(rhs) {}
-    CTxDestination(const CScriptID& rhs) : Super(rhs) {}
-
-    bool operator <  (const CTxDestination& _rhs) const {
-        const CTxDestination& lhs = *this;
-        const CTxDestination& rhs = _rhs;
-        return lhs < rhs;
-    }
-    bool operator == (const CTxDestination& _rhs) const {
-        const CTxDestination& lhs = *this;
-        const CTxDestination& rhs = _rhs;
-        return lhs == rhs;
-    }
-public:
-    base58string GetBase58address33() const;
-};
 
 const char* GetTxnOutputType(txnouttype t);
